pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "warlock-monitor"
version = "0.1.0"
description = "Linux-style CLI system monitor with Telegram alerts"
authors = [
    { name="Yaroslav", email="" }
]
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "psutil",
    "pyyaml",
    "python-telegram-bot>=20.0"
]

[project.scripts]
warlock = "warlock.cli:main"
